moved warnings and exceptions into hop.exceptions
